+++ codejail_service/settings/test.py
@@ -1,3 +1,7 @@
 """Settings for use during testing."""
 
 from codejail_service.settings.base import *  # pylint: disable=wildcard-import
+
+# We don't have any meaningful settings to add here for unit tests,
+# but codejail will still want the dictionary to exist.
+CODE_JAIL = {}